[WebGPU] Fix the 32-bit build https://bugs.webkit.org/show_bug.cgi?id=238053
Unreviewed.

* WebGPU/Buffer.mm:
(WebGPU::Buffer::getMappedRange):
(WebGPU::Buffer::mapAsync):


Canonical link: https://commits.webkit.org/248570@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@291447 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebGPU/WebGPU/Buffer.mm
@@ -213,7 +213,7 @@ void* Buffer::getMappedRange(size_t offset, size_t size)
     // FIXME: Use checked arithmetic.
     auto rangeSize = size;
     if (size == WGPU_WHOLE_MAP_SIZE)
-        rangeSize = std::max(static_cast<uint64_t>(0), m_size - static_cast<uint64_t>(offset));
+        rangeSize = static_cast<size_t>(std::max(static_cast<uint64_t>(0), m_size - static_cast<uint64_t>(offset)));
 
     // "If any of the following conditions are unsatisfied"
     if (!validateGetMappedRange(offset, rangeSize)) {
@@ -275,7 +275,7 @@ void Buffer::mapAsync(WGPUMapModeFlags mode, size_t offset, size_t size, Complet
     // FIXME: Use checked arithmetic.
     auto rangeSize = size;
     if (size == WGPU_WHOLE_MAP_SIZE)
-        rangeSize = std::max(static_cast<uint64_t>(0), static_cast<uint64_t>(m_size - offset));
+        rangeSize = static_cast<size_t>(std::max(static_cast<uint64_t>(0), static_cast<uint64_t>(m_size - offset)));
 
     // "If any of the following conditions are unsatisfied:"
     if (!validateMapAsync(mode, offset, rangeSize)) {
